+++ doc-examples/src/main/java/react4j/examples/arez/step2/Footer.java
@@ -7,7 +7,6 @@ import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.BaseContext;
 import react4j.core.BaseProps;
-import react4j.core.BaseState;
 import react4j.core.ReactNode;
 import react4j.dom.events.MouseEventHandler;
 import react4j.dom.proptypes.html.AnchorProps;
@@ -20,7 +19,7 @@ import static react4j.examples.arez.step2.Footer_.*;
 
 @ReactComponent
 class Footer
-  extends ReactArezComponent<BaseProps, BaseState, BaseContext>
+  extends ReactArezComponent<BaseProps, BaseContext>
 {
   @EventHandler( MouseEventHandler.class )
   void handleClearCompleted()