+++ doc-examples/src/main/java/react4j/examples/arez/step2/FooterTodoCount.java
@@ -6,7 +6,6 @@ import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.BaseContext;
 import react4j.core.BaseProps;
-import react4j.core.BaseState;
 import react4j.core.ReactNode;
 import react4j.dom.proptypes.html.HtmlProps;
 import react4j.examples.arez.step2.model.AppData;
@@ -15,7 +14,7 @@ import static react4j.examples.arez.step2.FooterTodoCount_.*;
 
 @ReactComponent
 class FooterTodoCount
-  extends ReactArezComponent<BaseProps, BaseState, BaseContext>
+  extends ReactArezComponent<BaseProps, BaseContext>
 {
   @Nonnull
   static ReactNode create()