+++ processor/src/test/resources/input/com/example/arez/AutorunArezReactComponent.java
@@ -6,12 +6,11 @@ import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.BaseContext;
 import react4j.core.BaseProps;
-import react4j.core.BaseState;
 import react4j.core.ReactNode;
 
 @ReactComponent
 class AutorunArezReactComponent
-  extends ReactArezComponent<BaseProps, BaseState, BaseContext>
+  extends ReactArezComponent<BaseProps, BaseContext>
 {
   @Autorun
   public void myAutorun()