+++ processor/src/test/resources/input/com/example/arez/BasicArezReactComponent.java
@@ -10,7 +10,7 @@ import react4j.core.ReactNode;
 
 @ReactComponent
 class BasicArezReactComponent
-  extends ReactArezComponent<BaseProps, BaseState, BaseContext>
+  extends ReactArezComponent<BaseProps, BaseContext>
 {
   @Nullable
   @Override