+++ processor/src/test/resources/input/com/example/arez/ComponentWithEventHandler.java
@@ -7,12 +7,11 @@ import react4j.annotations.ReactComponent;
 import react4j.arez.ReactArezComponent;
 import react4j.core.BaseContext;
 import react4j.core.BaseProps;
-import react4j.core.BaseState;
 import react4j.core.ReactNode;
 
 @ReactComponent
 class ComponentWithEventHandler
-  extends ReactArezComponent<BaseProps, BaseState, BaseContext>
+  extends ReactArezComponent<BaseProps, BaseContext>
 {
   @JsFunction
   public interface CustomHandler