+++ processor/src/test/resources/input/com/example/arez/ComponentWithNoAutoActionEventHandler.java
@@ -12,7 +12,7 @@ import react4j.core.ReactNode;
 
 @ReactComponent
 class ComponentWithNoAutoActionEventHandler
-  extends ReactArezComponent<BaseProps, BaseState, BaseContext>
+  extends ReactArezComponent<BaseProps, BaseContext>
 {
   @Nullable
   @Override