+++ processor/src/test/resources/input/com/example/arez/OverridingComponentDidUpdateComponent.java
@@ -11,7 +11,7 @@ import react4j.core.ReactNode;
 
 @ReactComponent
 class OverridingComponentDidUpdateComponent
-  extends ReactArezComponent<BaseProps, BaseState, BaseContext>
+  extends ReactArezComponent<BaseProps, BaseContext>
 {
   @Nullable
   @Override